+++ async-nats/src/jetstream/consumer/pull.rs
@@ -1370,6 +1370,10 @@ pub struct Config {
     /// to recover.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub durable_name: Option<String>,
+    /// A name of the consumer. Can be specified for both durable and ephemeral
+    /// consumers.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub name: Option<String>,
     /// A short description of the purpose of this consumer.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub description: Option<String>,
@@ -1418,6 +1422,9 @@ pub struct Config {
     /// Number of consumer replucas
     #[serde(default, skip_serializing_if = "is_default")]
     pub num_replicas: usize,
+    /// Force consumer to use memory storage.
+    #[serde(default, skip_serializing_if = "is_default")]
+    pub memory_storage: bool,
 }
 
 impl IntoConsumerConfig for &Config {
@@ -1430,6 +1437,7 @@ impl IntoConsumerConfig for Config {
     fn into_consumer_config(self) -> consumer::Config {
         jetstream::consumer::Config {
             deliver_subject: None,
+            name: self.name,
             durable_name: self.durable_name,
             description: self.description,
             deliver_group: None,
@@ -1450,6 +1458,7 @@ impl IntoConsumerConfig for Config {
             max_expires: self.max_expires,
             inactive_threshold: self.inactive_threshold,
             num_replicas: self.num_replicas,
+            memory_storage: self.memory_storage,
         }
     }
 }
@@ -1463,6 +1472,7 @@ impl FromConsumer for Config {
         }
         Ok(Config {
             durable_name: config.durable_name,
+            name: config.name,
             description: config.description,
             deliver_policy: config.deliver_policy,
             ack_policy: config.ack_policy,
@@ -1479,6 +1489,7 @@ impl FromConsumer for Config {
             max_expires: config.max_expires,
             inactive_threshold: config.inactive_threshold,
             num_replicas: config.num_replicas,
+            memory_storage: config.memory_storage,
         })
     }
 }